+++ usr/lib/cca_stdll/cca_stdll.mk
@@ -35,7 +35,8 @@ opencryptoki_stdll_libpkcs11_cca_la_SOURCES =				\
 	usr/lib/common/mech_ssl3.c usr/lib/common/verify_mgr.c		\
 	usr/lib/common/p11util.c usr/lib/common/sw_crypt.c		\
 	usr/lib/common/shared_memory.c usr/lib/common/profile_obj.c	\
-	usr/lib/cca_stdll/cca_specific.c usr/lib/common/attributes.c
+	usr/lib/cca_stdll/cca_specific.c usr/lib/common/attributes.c	\
+	usr/lib/common/dlist.c
 
 if ENABLE_LOCKS
 opencryptoki_stdll_libpkcs11_cca_la_SOURCES +=				\
