+++ usr/lib/ep11_stdll/ep11_stdll.mk
@@ -36,7 +36,7 @@ opencryptoki_stdll_libpkcs11_ep11_la_SOURCES =				\
 	usr/lib/common/utility.c usr/lib/common/trace.c			\
 	usr/lib/common/mech_list.c usr/lib/common/shared_memory.c	\
 	usr/lib/common/attributes.c usr/lib/common/sw_crypt.c		\
-	usr/lib/common/profile_obj.c					\
+	usr/lib/common/profile_obj.c usr/lib/common/dlist.c		\
 	usr/lib/common/pkey_utils.c					\
 	usr/lib/ep11_stdll/new_host.c usr/lib/ep11_stdll/ep11_specific.c
