+++ usr/lib/soft_stdll/soft_stdll.mk
@@ -32,7 +32,8 @@ opencryptoki_stdll_libpkcs11_sw_la_SOURCES =				\
 	usr/lib/common/utility.c usr/lib/common/verify_mgr.c		\
 	usr/lib/common/trace.c usr/lib/common/mech_list.c		\
 	usr/lib/common/shared_memory.c usr/lib/common/profile_obj.c	\
-	usr/lib/soft_stdll/soft_specific.c usr/lib/common/attributes.c
+	usr/lib/soft_stdll/soft_specific.c usr/lib/common/attributes.c	\
+	usr/lib/common/dlist.c
 
 if ENABLE_LOCKS
 opencryptoki_stdll_libpkcs11_sw_la_SOURCES +=				\
