+++ usr/sbin/pkcsslotd/pkcsslotd.mk
@@ -21,5 +21,6 @@ usr_sbin_pkcsslotd_pkcsslotd_SOURCES =							\
 	usr/sbin/pkcsslotd/socket_server.c
 
 nodist_usr_sbin_pkcsslotd_pkcsslotd_SOURCES = \
-	usr/lib/common/parser.h usr/lib/common/parser.c	usr/lib/common/lexer.c
+	usr/lib/common/parser.h usr/lib/common/parser.c	usr/lib/common/lexer.c		\
+	usr/lib/common/dlist.c
 usr/sbin/pkcsslotd/slotmgr.$(OBJEXT): usr/lib/common/parser.h
